Referencify tuples
I'm not sure "initialized" is the correct word, but I feel like our
verbiage around place and value expressions need to be cleaned up
anyways. I originally went with "moved", but that's not really correct.
I think I want to call it a "value operand", but I need to discuss with
others on that first. In any case, it's a lot more correct than what
was there before.

I renamed `unit_x` to `point` in an example so as to not cause confusion
with the unit tuple.