Compound operator expressions
I don't actually know if the syntactic sugar is correct, so hopefully
the lang team can sign off that it is in-deed correct.

I also don't know if I can link to the test files without linking to
Github, so I'm linking to Github. We could probably make linkcheck
work for these eventually, if needed?