Attempt to explain that &mut cannot alias
This is a clumsy attempt at making the "`&mut` is an exclusive
reference" part of the aliasing rule more discoverable. I feel like this
is an inherent property of the type so I was surprised to not find that
information here. This is echoing parts of the docs for [UnsafeCell] and
[ptr::as_mut], so another place where a newcomer could potentially get
this info.

I think it's easy for learners to assume that `&mut` can alias since
many popular languages have comparatively lax aliasing rules.

[UnsafeCell]: https://doc.rust-lang.org/stable/std/cell/struct.UnsafeCell.html
[ptr::as_mut]: https://doc.rust-lang.org/std/primitive.pointer.html#method.as_mut-1