Add an anchor to the "forwarding macro fragments" paragraph
Since this is a subtle idiosyncracy into which macro users sometimes bump, it can be incredibly help ful to give them a pin-pointed link to the exact paragraph talking about this. Hence the idea of adding some kind of header, and thus an anchor, to it. Since it's just a paragraph, I've gone for a very low header, such as h4.

cc @nilstrieb