Remove "expression for a match arm" wording
It's unclear what this bullet means. If it means a match guard, that is covered in the line above. If it referred to the scrutinee, that would be wrong. I don't think it's possible to embed an expression inside of a match arm _pattern_.

This came up in the [lang team triage meeting][meeting].

[meeting]: https://hackmd.io/JACzq-ssT4qtH0DS3B8nQQ?both#%E2%80%9CIntroduce-terminating-scope-for-tail-expressions-of-breakable-scopes%E2%80%9D-rust106493