type-layout: be more specific about 32-bit alignments
The rust-reference implies that 64-bit types are aligned to 32-bit for
platforms with 32-bit addresses. This is not necessarily correct. Fix
the wording.

Note that there is no general rule how data-types greater than the
native address size are aligned. On most Unix'y systems, they use the
native alignment of the platform. However, the Windows ABI aligns them
to their size (up to at least 64-bit).

There are advantages for either of those decisions. But we should at
least make clear that there is no fixed rule for 32-bit platforms.

Signed-off-by: David Rheinsberg <david@readahead.eu>