Add tracking of rule prefixes
This adds tracking of rule prefixes (like `asm.ts-args` is an interior
prefix of `asm.ts-args.syntax`). This will be used for test linking,
to check for tests that link to these.

I'm not sure what we'll want to do with these long-term, whether we
want to allow linking these "uber" rules, or remove them, or just
warn about them.