Add test linking
This adds the ability to link rust-lang/rust tests to reference rule
annotations. Rules now have a link that pops up a list of tests that
are associated with that rule. Additionally, there is a new appendix
that lists the number of rules in each chapter, how many tests are
associated, and various summaries.

This requires a local checkout of rust-lang/rust which is pointed to
by the `SPEC_RUST_ROOT` environment variable.