Some more elaboration on do_not_recommend
This adds some more explanation about the use of
`diagnostic::do_not_recommend`, with the intent of helping the reader to
understand how to use it. This also breaks that discussion out into a
note.

This also adds a before/after to the example to better show how it
changes.