Move heading-related rules before headings
i.e. this replaces
```markdown
# UwU

r[uwu]
```
with
```markdown
r[uwu]
# UwU
```

This is required because the new grid system assumes that a rule is
related to the *following* element. But also, it makes it clear in the
source which element the rule is related to, as it's curently ambiguous.
Also (2), existing annotations are not 100% consistent.

I've done this using the following two irregular "regex" replacements:
1. `([^\]]\n|^(?<!\n))(^#+ [^\n]+)\n\n?(r\[.+\])\n*(r\[)`
   ==>> `$1$3\n$2\n\n$4` (~255 cases)
2. `([^\]]\n|^(?<!\n))(^#+ [^\n]+)\n\n?(r\[.+\])\n+`
   ==>> `$1$3\n$2\n\n` (~99 cases)

With the latter one being reviewed manually, since it can produce bad
replacements in cases like this, where the rule relates to the following
paragraph/element, rather than the preceding heading:
```markdown
# OwO

r[meow]
>////< mEeoowW prRRrr purrrrrrRr :3 mrrrrrreww meEooww purrrrrr~ Mrrrew
```