Drop sentence about UB of how to violate the Rust runtime
I felt like this sentence is making an over-bold statement. There are
ways to violate the assumptions of the Rust runtime using Rust code.
Usually that is only with `unsafe` code that is otherwise violating
requirements, though there may still be some holes where it can be done
in safe code (hopefully those holes continue to be plugged). I feel more
comfortable without this, since the primary sentence (don't violate the
runtime!) is enough on its own.