Normatively define the abort and unwind runtimes
We specifically talk about `abort` and `unwind` already. Let's just go
ahead and up-front define that there are two, and not be so hand-wavey
about it.

In theory the standard library could introduce additional runtimes, and
it doesn't have a specific impact on the language per se. However, this
is a blurred line where I think it is better to be clear about what
there is. We can always revisit this later and adjust if things change.