Rework range pattern rules
I was not particularly happy with how these were being presented, as I
found it a bit confusing how it was combining some of the ranges, and
some of the terminology it was using.

I've shifted it so that it describes each of the five range kinds
separately, using the same terminology from std (and the grammar).

This also just has some slight wording tweaks for clarity.