Revise `asm_goto` changes
Let's make some clarifications of substance and editorial tweaks to
the changes for `asm_goto`.  We add an example, we change various
wording to be more clear -- in particular, we significantly revise the
section on `noreturn` to unify the statement of the rules -- and we
replace remaining uses of "asm code" with "assembly code" to match our
verbiage elsewhere.

The main clarification of substance is that for requirements that must
be upheld when falling through the assembly, such as he requirement to
restore the stack pointer, we must also note that the requirement must
be satisfied before jumping to any `label` blocks.