Add the ability for rules to be specified in link definitions
This adds the ability to link to rules from a link definition. For example:

```markdown
See [this rule].

[this rule]: expr.array
```

This will convert `[this rule]` to point to the link of the file for
that rule.

This is somewhat hacky, as parsing markdown with regex is unwise (and
this is intentionally incomplete), and could maybe be done more
efficiently.