Clarify all borrow operators in the precedence table
This removes `&` and `&mut` and instead just links to the full grammar
for what the borrow operators are. There are many other variants, like
`&&` or `&raw`, and I didn't want to write all of them down. The
keywords after the `&&` are just modifiers, and are not really separate
things to be considered anyways.

Closes https://github.com/rust-lang/reference/issues/1726