Rename IsolatedCR to CR
This renames IsolatedCR to CR. I felt like it wasn't exactly necessary
since we have rewritten things so that it is clear that there is an
input transformation which resolves this (`input.crlf`). We also never
really defined what it meant.

I also felt like there was room for confusion. For example, an input
containing `CR CR LF LF` would get normalized to `CR LF`. The `CR` there
is not isolated.