Rewrite how keywords are listed
This changes the keyword listings so that they are just lists instead of
lexer rules. We never used the named rules, and I don't foresee us ever
doing that. Also, the `IDENTIFIER_OR_KEYWORD` rule meant that we never
needed to explicitly identify these keywords as lexer tokens.

This helps avoid problems when building the grammar graph for missing
connections.