Add a lexer rule for punctuation
This is intended to help define what a "token" is via the grammar (and
to fill a missing hole in our token definition).

I waffled on how to define delimiters, whether they should be separate
somehow. In practice I think it should be fine to clump them all
together. This mainly only matters for TokenTree which already excludes
the delimiters.