Mark grammar roots inline in productions
We check that the list of grammar "roots" -- that is, productions that
are not used in any other production -- is what we expect it to be.

We had hard coded this list of roots in `mdbook-spec`.  Let's instead
add a way to specify this in our syntax for productions by prefixing
the production with `@root`.