Cross link between grammar and railroad diagrams
When reviewing a production in the grammar, one often wants to quickly
find the corresponding railroad diagram, and when reviewing a railroad
diagram, one often wants to quickly find the corresponding production
in the grammar.

Let's make this easy by linking each production in the grammar to the
corresponding railroad diagram, and from the name of each railroad
diagram to the corresponding production in the grammar.

When clicking on a production in the grammar, we'll automatically
display the railroad diagrams if those are not already displayed.