Fix check for new and unexpected roots
We track the "roots" in our grammar -- those productions that aren't
used in any other production.  We want to report when a new root
appears or when something that's expected to be a root no longer is
one.  However, we were reporting the latter case as the former instead
of reporting it separately as intended.  Let's fix that.