Remove broken footnote links from grammar summary
Our grammar syntax supports linking to footnotes which we render in
the Markdown version.  This works when the rendering is done on the
same page that the footnote appears.

On the grammar summary page, however, where we also render the
grammar, these footnotes are not present, and so the links were all
broken.

Let's fix this for now by not rendering these footnote links on the
grammar summary page.