Refactor rendering with `RenderCtx`
We thread certain contextual information through to most of our rendering
functions.  Currently this information includes whether the rendering
is for the summary page and the two link maps.  Adding new information
to thread through, or changing something that was only using a subset
to use more of the context, requires touching a lot of places.

Let's improve this by refactoring these details into a `RenderCtx`
that we'll thread through everywhere instead.  In this way, new
details can simply be added to this struct, and callees can use
whatever details are needed.