Remove normative cites to unsafe code guidelines doc
The unsafe code guidelines document describes itself currently as
largely abandoned and indicates that those points that have found
consensus are documented in the Reference.  Let's not then normatively
cite that other document.

Let's also remove a line about "these rules do not apply..." where
"these" changed meaning when this line was copied from elsewhere.  In
the original, "these" meant rules about `readonly` and `nomem`,
whereas here, it would have seemed to refer to the rules in the UCG,
and that wouldn't make sense.