Clarify rules on callee/caller saved registers
For naked functions, let's break out the rules for callee and caller
saved registers and name the rule identifiers more clearly.

For the caller saved registers, let's drop the caveat about "even if
they are not used for the return value", as it doesn't really add
anything to the rule.