Remove StructExprTuple and StructExprUnit
This removes the grammar rules StructExprTuple and StructExprUnit, and
removes these as distinct expressions. Instead, a note block is used to
let the reader know how the constructors can be accessed in the value
namespace.

This stems back to the beginning of this documentation, which presumably
was presenting these as distinct kinds of expressions just as a
simplification or to match what people's mental models might be.
However, they are not distinct expressions, and I think it is misleading
to pretend that they are.

Closes https://github.com/rust-lang/reference/issues/1802