Update let chain grammar for minimum precedence
This adds some more exclusions to the let chain grammar to capture the
minimum precedence while parsing chains. Note that this is incomplete,
see https://github.com/rust-lang/reference/issues/1811. I didn't want to
block things on finalizing this.

I also reworked the `while` grammar to just reuse the grammar from `if`
to avoid the duplication.