document that the `rustc` tool namespace is reserved
this section in the reference is not quite right; or at least is an
incomplete summary. here is a more detailed explanation of what's
currently implemented:

|Tool|Lints|Attributes|
|-|-|-|
|`clippy`|✅|✅|
|`rustfmt`|❌|✅|
|`miri`|❌|✅|
|`rust_analyzer`|❌|✅|
|`rustdoc`|✅|❌|
|`rustc`|✅ (with `-Z unstable-options`)|❌|
|`diagnostic`|❌|✅|

i'm not sure how to write that up, and it doesn't belong in the
attributes section anyway, so i've just added this short snippet.