Align pattern destructuring with rest of doc
Previously, `patterns.destructure.placeholder` referred to the `_` in a
destructuring pattern as "placeholder" and `..` as "wildcard". This
directly contradicted the definition of struct patterns and tuple
(struct) patterns, which (indirectly) define `_` as a wildcard pattern
and `..` as a structure pattern et cetera (for structs, enums, and union
values) or rest pattern (for tuples, tuple structs, slices, and arrays).

Update the paragraph to use "wildcard pattern", "et cetera", and "rest
pattern" instead.  Moreover, add links to their respective
sections/statements. Finally, update the rule name to
`r[patterns.destructure.wildcard]`.