Fix recursive root-accessible grammar check
This fixes a bug in how the `@root` grammar validation was done. It was
not properly handling recursive rules because when walking the tree it
would find itself, and thus mark it as "used". What we really want is to
only walk starting from the root set, and don't remove recursive
definitions from the root set.