Fix inconsistent heading depth
This fixes several pages that did not start with a level one heading.

The unsafe page also had a weird use of level five headings. It's not
entirely clear to me why. The page originally used a list, so I have
switched it back to that. If in the future we decide to add more content
to one of these items, it can be broken out into a separate section like
integer overflow did.