Fix smart punctuation inside grammar terminals
Smart punctuation was converting `...` to `…`. This fixes it so that it
will be escaped to avoid the conversion. This normally doesn't happen
inside a backtick code block, but since we are using spans for
terminals, pulldown-cmark is accidentally picking them up.

This also includes hyphen to handle en-dash and em-dash, though we don't
have those.

Fixes https://github.com/rust-lang/reference/issues/1867