Add clarifications
In the text as revised, it wasn't clear what rule prevented a mutable
reference from appearing in the final value of a constant item, so
let's add a rule specifically for that.

Similarly, it may not be immediately clear what rule prevents a
reference to a lifetime-extended interior mutable temporary created in
the initializer, so let's add an admonition section describing how
other rules apply to prevent this.