Revise, clarify, and add examples
Rather than talking about lifetime-extended temporaries in the
top-level scope of an initializer, which is maybe a bit ambiguous,
let's speak directly to the result of the lifetime extension, which is
that these temporaries disallowed for borrows would have their
lifetimes extended to the end of the program.

Let's also speak about place expressions, rather than places, as
that's more precise here.

We'll add examples throughout.  Thanks to RalfJ for the substance of
many of these.