Remove outdated comment about non-copy unions
I'm not sure when exactly this was removed, but non-copy union fields
aren't allowed, and there is no nightly feature to allow them AFAIK.

There was various discussion in the tracking issue
https://github.com/rust-lang/rust/issues/55149 and various changes like
https://github.com/rust-lang/rust/pull/97995 on what is allowed, but I
did not dig down to find the exact change that closed the door.

The allowed field kinds are described in items.union.field-restrictions.

Also, in general the reference shouldn't be referring to unstable
features.