Revise editorially
In most of the examples for these attributes, there's no lead-in, so
let's remove the lead-in here.

Let's also say "form" rather than "thing", even though "thing" does
have a certain appeal to it.

In our rule for duplicates, let's add a forward reference to the
exception to that.

Finally, we say that the attribute applied at the crate level does not
remove the crate; later, for an example, let's clarify that it only
removes the content of the crate.

Most or all of this was preexisting to this PR.