Document (tuple) struct pattern namespace behavior
We had not documented what namespaces were used when matching struct
and tuple struct patterns.  As it turns out, this is a bit nuanced.
Let's document the current behavior, describe the nature of this
nuance with an example, and note the possible uncertainly about the
desirability of this behavior long-term.