Rename question mark operator
This renames the "question mark operator" section to the "try
propagation expression", and also drops the grammar name "error
propagation expression".

The "error propagation expression" isn't really accurate, since `?`
isn't just for errors.

The intent is to describe the expression by what it does (as most other
expressions do) instead of how it looks. It also felt a little awkward
to call it the "question mark expression".